mim_startex/requirements.txt
camelot-py==1.0.9
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
et-xmlfile==2.0.0
numpy==2.3.3
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
pandas==2.3.3
pdfminer-six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.23
pypdf==5.9.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tabulate==0.9.0
tzdata==2025.2
